feat: separated types from executable code and modifying the build.mjs file accordingly
