+++ src/api/debug.ts
@@ -1,10 +1,5 @@
 import { sendMessage } from '../ws/websocket';
-
-export enum LoggerType {
-  WARNING = 'WARNING',
-  ERROR = 'ERROR',
-  INFO = 'INFO'
-};
+import type { LoggerType } from '../../types/enums';
 
 export function log(message: string, type?: LoggerType): Promise<void> {
     return sendMessage('debug.log', { message, type });
