+++ src/api/extensions.ts
@@ -1,9 +1,5 @@
 import * as websocket from '../ws/websocket';
-
-export interface ExtensionStats {
-    loaded: string[];
-    connected: string[];
-}
+import type { ExtensionStats } from '../../types/api/extensions';
 
 export function dispatch(extensionId: string, event: string, data?: any): Promise<void> {
     return new Promise(async (resolve: any, reject: any) => {
