+++ types/api/app.d.ts
@@ -1,0 +1,7 @@
+export interface OpenActionOptions {
+    url: string;
+}
+
+export interface RestartOptions {
+    args: string;
+}
