+++ types/api/extensions.d.ts
@@ -1,0 +1,4 @@
+export interface ExtensionStats {
+    loaded: string[];
+    connected: string[];
+}
