+++ types/api/init.d.ts
@@ -1,0 +1,3 @@
+export interface InitOptions {
+    exportCustomMethods?: boolean;
+}
