+++ types/api/updater.d.ts
@@ -1,0 +1,5 @@
+export interface Manifest {
+    applicationId: string;
+    version: string;
+    resourcesURL: string;
+}
